Fix parsing of new blocks in state
This change fixes the parsing of a Terraform state by only creating new blocks if the line starts with `# ` - note the space after the hash.

This is required because the state can contain arbitrary data, e.g. the raw content of a file (from `data.local_file`, which can start with a `#` - this is what happened in my case.

I hope that hash + whitespace will only occur if a new block starts.